# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Data Process Service",
        "title": "Data Process API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/csv/process": {
            "post": {
                "description": "Insert / Update Process CSV",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "csv"
                ],
                "summary": "Process CSV",
                "parameters": [
                    {
                        "description": "Array Path CSV",
                        "name": "csv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handler.ProcessCSVRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handler.ProcessCSVRequest": {
            "type": "object",
            "required": [
                "file_paths"
            ],
            "properties": {
                "file_paths": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}